+++ rest/messaging/v1/model_messaging_v1_us_app_to_person.go
@@ -68,4 +68,6 @@ type MessagingV1UsAppToPerson struct {
 	Url *string `json:"url,omitempty"`
 	// A boolean that specifies whether campaign is a mock or not. Mock campaigns will be automatically created if using a mock brand. Mock campaigns should only be used for testing purposes.
 	Mock *bool `json:"mock,omitempty"`
+	// Details indicating why a campaign registration failed. These errors can indicate one or more fields that were incorrect or did not meet review requirements.
+	Errors *[]interface{} `json:"errors,omitempty"`
 }