+++ twilio.go
@@ -30,6 +30,7 @@ import (
 	FlexV2 "github.com/twilio/twilio-go/rest/flex/v2"
 	FrontlineV1 "github.com/twilio/twilio-go/rest/frontline/v1"
 	InsightsV1 "github.com/twilio/twilio-go/rest/insights/v1"
+	IntelligenceV2 "github.com/twilio/twilio-go/rest/intelligence/v2"
 	IpMessagingV1 "github.com/twilio/twilio-go/rest/ip_messaging/v1"
 	IpMessagingV2 "github.com/twilio/twilio-go/rest/ip_messaging/v2"
 	LookupsV1 "github.com/twilio/twilio-go/rest/lookups/v1"
@@ -77,6 +78,7 @@ type RestClient struct {
 	FlexV2          *FlexV2.ApiService
 	FrontlineV1     *FrontlineV1.ApiService
 	InsightsV1      *InsightsV1.ApiService
+	IntelligenceV2  *IntelligenceV2.ApiService
 	IpMessagingV1   *IpMessagingV1.ApiService
 	IpMessagingV2   *IpMessagingV2.ApiService
 	LookupsV1       *LookupsV1.ApiService
@@ -171,6 +173,7 @@ func NewRestClientWithParams(params ClientParams) *RestClient {
 	c.FlexV2 = FlexV2.NewApiService(c.RequestHandler)
 	c.FrontlineV1 = FrontlineV1.NewApiService(c.RequestHandler)
 	c.InsightsV1 = InsightsV1.NewApiService(c.RequestHandler)
+	c.IntelligenceV2 = IntelligenceV2.NewApiService(c.RequestHandler)
 	c.IpMessagingV1 = IpMessagingV1.NewApiService(c.RequestHandler)
 	c.IpMessagingV2 = IpMessagingV2.NewApiService(c.RequestHandler)
 	c.LookupsV1 = LookupsV1.NewApiService(c.RequestHandler)